--- a/.travis.yml
+++ b/.travis.yml
@@ -1,7 +1,7 @@
 language: rust
 sudo: false
 rust:
-- 1.2.0
+- stable
 - beta
 - nightly
 script: