@jpastuszek 2015-11-01
Ups, forgot to rebase/squash... I will create another PR if we want this.