@aldanor 2015-11-03
Hey, thanks, a lot of stuff to read through, but I prob won't be able to get through it till the weekend most def(because of the web summit and all...)